+++ app/lib/data.ts
@@ -12,16 +12,65 @@ import {
 
 
 
-export async function fetchLighthouseRanges(
-  
-) {
+export async function fetchLighthouseRanges() {
   try {
     const data = await sql<LighthouseType>
     `SELECT 
             lighthouse.id, 
             lighthouse.name, 
             lighthouse.range
-      FROM lighthouse`;
+      FROM lighthouse
+      ORDER BY lighthouse.range`;
+    return data.rows;
+  } catch (error) {
+    console.error('Database Error:', error);
+    throw new Error('Failed to fetch lighthouse location data.');
+  }
+}
+
+export async function fetchLighthouseABWMetrics() {
+  try {
+    const data = await sql<LighthouseType>
+    `SELECT 
+            lighthouse.id, 
+            lighthouse.name, 
+            lighthouse.abovewater
+      FROM lighthouse
+      ORDER BY lighthouse.abovewater`;
+    return data.rows;
+  } catch (error) {
+    console.error('Database Error:', error);
+    throw new Error('Failed to fetch lighthouse location data.');
+  }
+}
+
+export async function fetchLighthouseTowerMetrics() {
+  try {
+    const data = await sql<LighthouseType>
+    `SELECT 
+            lighthouse.id, 
+            lighthouse.name, 
+            lighthouse.towerheight
+      FROM lighthouse
+      ORDER BY lighthouse.towerheight`;
+    return data.rows;
+  } catch (error) {
+    console.error('Database Error:', error);
+    throw new Error('Failed to fetch lighthouse location data.');
+  }
+}
+
+export async function fetchLighthouseAges() {
+  try {
+    const data = await sql<LighthouseType>
+    `SELECT 
+            lighthouse.id, 
+            lighthouse.name, 
+            lighthouse.constructed,
+            lighthouse.currentdate,
+            (lighthouse.currentdate - lighthouse.constructed)/365 AS "age"
+      FROM lighthouse
+      ORDER BY age`;
     return data.rows;
   } catch (error) {
     console.error('Database Error:', error);
@@ -44,13 +93,13 @@ export async function fetchLighthouses(
             lighthouse.name, 
             lighthouse.latitude, 
             lighthouse.longitude,
-            lighthouse.aboveWater,
-            lighthouse.towerHeight, 
+            lighthouse.abovewater,
+            lighthouse.towerheight, 
             lighthouse.range, 
-            lighthouse.greatLighthouse,
+            lighthouse.greatlighthouse,
             lighthouse.constructed,
-            lighthouse.currentDate,
-            (lighthouse.currentDate - lighthouse.constructed)/365 AS "age",
+            lighthouse.currentdate,
+            (lighthouse.currentdate - lighthouse.constructed)/365 AS "age",
             lighthouse.image_url
      FROM lighthouse
      WHERE