+++ app/ui/home/aboveWaterChart.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import { lighthouses } from  "@/app/lib/placeholder-data";
-
 import {
     CartesianGrid,
     XAxis,
@@ -13,35 +11,26 @@ import {
     ResponsiveContainer,
   } from "recharts";
 
+type LighthouseABWProps = {
+  abwData: {
+    id: string;
+    name: string;
+    abovewater: number;
+  }[]
+}
 
-// type LighthouseCharacterProps = {
-//   characters: {
-//     id: string;
-//     name: string;
-//     latitude: number;
-//     longitude: number;
-//     aboveWater:number;
-//     towerHeight:number;
-//     range:number;
-//     age:number;
-//   }[]
-// }
-
-export default function AboveWaterGraph() {
-  const aboveData = [...lighthouses]
-  const sortedData = aboveData.sort((a,b) => a.aboveWater - b.aboveWater);
-  // console.log("above...", sortedData);
+export default function AboveWaterGraph({abwData} : LighthouseABWProps) {
         return (
           <>
             <ResponsiveContainer width="100%" minHeight={180}>
-              <BarChart data={sortedData}
+              <BarChart data={abwData}
                   margin={{ top: 5, right: 30, left: 20, bottom: 5 }}>
                   <CartesianGrid strokeDasharray="3 3" />
                   <XAxis dataKey="name" />
                   <YAxis />
                   <Tooltip />
                   <Legend />
-                  <Bar dataKey="aboveWater" stroke="#8884d8"/>
+                  <Bar dataKey="abovewater" stroke="#2284d8"/>
               </BarChart>
             </ResponsiveContainer>
           </>