+++ app/ui/home/sidenav.tsx
@@ -1,6 +1,7 @@
 import Link from 'next/link';
 import NavLinks from '@/app/ui/home/nav-links';
 import BeaconLogo from '@/app/ui/ibw-logo';
+import Footer from '../footer';
 
 
 export default function SideNav() {
@@ -12,6 +13,7 @@ export default function SideNav() {
       <div className="flex grow flex-row justify-between space-x-2 md:flex-col md:space-x-0 md:space-y-2">
         <NavLinks />
         <div className="hidden h-auto w-full grow rounded-md bg-gray-50 md:block"></div>
+        <Footer/>
       </div>
     </div>
   );