+++ app/ui/home/towerChart.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import { lighthouses } from  "@/app/lib/placeholder-data";
-
 import {
     CartesianGrid,
     XAxis,
@@ -14,35 +12,26 @@ import {
   } from "recharts";
 
 
-// type LighthouseCharacterProps = {
-//   characters: {
-//     id: string;
-//     name: string;
-//     latitude: number;
-//     longitude: number;
-//     aboveWater:number;
-//     towerHeight:number;
-//     range:number;
-//     age:number;
-//   }[]
-// }
+type LighthouseTowerProps = {
+  towerData: {
+    id: string;
+    name: string;
+    towerheight:number;
+  }[]
+}
 
-
-export default function TowerGraph() {
-  const towerData = [...lighthouses]
-  const sortedData = towerData.sort((a,b) => a.towerHeight - b.towerHeight);
-  // console.log("tower...", sortedData);
+export default function TowerGraph({towerData} : LighthouseTowerProps) {
         return (
           <>
             <ResponsiveContainer width="100%" minHeight={180}>
-              <BarChart data={sortedData}
+              <BarChart data={towerData}
                   margin={{ top: 5, right: 30, left: 20, bottom: 5 }}>
                   <CartesianGrid strokeDasharray="3 3" />
                   <XAxis dataKey="name" />
                   <YAxis />
                   <Tooltip />
                   <Legend />
-                  <Bar dataKey="towerHeight" stroke="#8884d8" />
+                  <Bar dataKey="towerheight" stroke="#4484d8" />
               </BarChart>
             </ResponsiveContainer>
           </>